chopper/requirements.txt
loguru
ffmpeg-python
fastapi
uvicorn
python-multipart
